# server/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /tables:
    get:
      operationId: TablesController_findAll
      parameters:
        - name: status
          required: false
          in: query
          description: Filter by table status
          schema:
            enum:
              - available
              - occupied
              - reserved
              - cleaning
            type: string
        - name: includeDeleted
          required: false
          in: query
          description: Include deleted tables (true/false)
          schema:
            type: string
      responses:
        '200':
          description: Returns list of tables
      summary: Get all tables
      tags: &ref_5
        - tables
    post:
      operationId: TablesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTableDto'
      responses:
        '201':
          description: Table created successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
      security:
        - JWT-auth: []
      summary: Create new table
      tags: *ref_5
  /tables/{id}:
    get:
      operationId: TablesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Table ID
          schema:
            type: string
        - name: includeDeleted
          required: false
          in: query
          description: Include deleted tables (true/false)
          schema:
            type: string
      responses:
        '200':
          description: Returns table details
        '404':
          description: Table not found
      security:
        - JWT-auth: []
      summary: Get table by ID
      tags: *ref_5
    put:
      operationId: TablesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Table ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTableDto'
      responses:
        '200':
          description: Table updated successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Table not found
      security:
        - JWT-auth: []
      summary: Update table
      tags: *ref_5
    delete:
      operationId: TablesController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Table ID
          schema:
            type: string
      responses:
        '200':
          description: Table deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Table not found
      security:
        - JWT-auth: []
      summary: Delete table (soft delete)
      tags: *ref_5
  /tables/{id}/restore:
    post:
      operationId: TablesController_restore
      parameters:
        - name: id
          required: true
          in: path
          description: Table ID
          schema:
            type: string
      responses:
        '200':
          description: Table restored successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Table not found
      security:
        - JWT-auth: []
      summary: Restore deleted table
      tags: *ref_5
  /tables/{id}/status:
    patch:
      operationId: TablesController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          description: Table ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTableStatusDto'
      responses:
        '200':
          description: Table status updated successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Table not found
      security:
        - JWT-auth: []
      summary: Update table status
      tags: *ref_5
  /tables/{id}/qr-code:
    get:
      operationId: TablesController_generateQrCode
      parameters:
        - name: id
          required: true
          in: path
          description: Table ID
          schema:
            type: string
      responses:
        '200':
          description: Returns QR code data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrCodeResponseDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Table not found
      security:
        - JWT-auth: []
      summary: Generate QR code for table
      tags: *ref_5
...
  /customer/tables/{id}:
    get:
      operationId: CustomerController_getTable
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Return the table information.
        '404':
          description: Table not found.
      summary: Get table information
      tags: *ref_9
  /customer/tables/{id}/qr-code:
    get:
      operationId: CustomerController_getTableQrCode
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Return the QR code for the table.
        '404':
          description: Table not found.
      summary: Get table QR code
      tags: *ref_9
...
info:
  title: Restaurant Management API
  description: |2-

        # Restaurant Management System API Documentation
        
        Hệ thống quản lý nhà hàng toàn diện với các tính năng đầy đủ cho việc vận hành nhà hàng hiện đại.
        
        ## Tính năng chính
        - **🔐 Xác thực & Phân quyền**: JWT-based authentication với phân quyền theo vai trò
        - **👥 Quản lý người dùng**: Admin, Staff, Kitchen, Waiter, Warehouse roles
        - **🍽️ Quản lý thực đơn**: Categories, dishes, menus với theo dõi nguyên liệu
        - **📋 Quản lý đơn hàng**: Đặt bàn qua QR, workflow bếp, theo dõi trạng thái
        - **📦 Quản lý kho**: Theo dõi nguyên liệu, quản lý lô hàng, nhập/xuất kho
        - **💳 Thanh toán**: Tích hợp VNPay cho thanh toán online
        - **🪑 Quản lý bàn**: Tạo QR code, theo dõi trạng thái bàn
        - **📊 Báo cáo**: Báo cáo tài chính, báo cáo kho
        
        ## Xác thực API
        Hầu hết các endpoint yêu cầu xác thực qua JWT token. Thêm token vào header Authorization:
        `Authorization: Bearer <your-jwt-token>`
        
        ## Phân quyền người dùng
        - **ADMIN**: Toàn quyền hệ thống
        - **STAFF**: Các thao tác nhân viên phục vụ
        - **KITCHEN**: Các thao tác đặc thù bếp
        - **WAITER**: Quản lý bàn và đơn hàng
        - **WAREHOUSE**: Quản lý kho
        
        ## HTTP Status Codes
        API trả về các mã trạng thái HTTP chuẩn:
        - **200**: Thành công
        - **201**: Tạo mới thành công
...
servers:
...
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Nhập JWT token (không cần prefix Bearer)
      in: header
  schemas:
...
    CreateTableDto:
      type: object
      properties:
        name:
          type: string
          description: Table name
          example: Bàn 1
        capacity:
          type: number
          description: Table capacity (number of seats)
          example: 4
          minimum: 1
          maximum: 50
      required:
        - name
        - capacity
    UpdateTableDto:
      type: object
      properties:
        name:
          type: string
          description: Table name
          example: Bàn 1
        capacity:
          type: number
          description: Table capacity (number of seats)
          example: 4
          minimum: 1
          maximum: 50
        status:
          type: string
          description: Table status
          enum:
            - available
            - occupied
            - reserved
            - cleaning
    UpdateTableStatusDto:
      type: object
      properties:
        status:
          type: string
          description: Table status
          enum:
            - available
            - occupied
            - reserved
            - cleaning
          example: available
      required:
        - status
    QrCodeResponseDto:
      type: object
      properties:
        qrCode:
          type: string
          description: The QR code as base64 encoded string
          example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
        url:
          type: string
          description: The URL that the QR code points to
          example: >-
            http://localhost:3000/customer/menu?tableId=123e4567-e89b-12d3-a456-426614174000
        table:
          type: object
          description: Table information
          example:
            id: 123e4567-e89b-12d3-a456-426614174000
            name: Bàn 1
            capacity: 4
      required:
        - qrCode
        - url
        - table
...